+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Change log
 -----------
 
+* Add configurable Docker registry for resinhup [Michal]
 * Switch to the new device register endpoint which exchanges the provisioning key for a device api key [Page]
 * Switch to using a device api key for api calls and vpn authentication if it is present